requirements.txt
gpiozero
luma.core
luma.led-matrix
luma.oled
Pillow
numpy
# lgpio is installed system-wide via python3-lgpio
